Server logic has been moved to the server folder
