+++ src/server/router/router.ts
@@ -1,0 +1,8 @@
+import express from 'express';
+import serverRenderMiddleware from "@/server/server-render-middleware";
+
+const router = express.Router();
+
+router.get('/*', serverRenderMiddleware);
+
+export default router;
